// React/app/GerAdv_TS/Fase/Interfaces/interface.Fase.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
                     // copyright © 2000-2025 Menphis - Sistemas Inteligentes
                     // This file is part of the Source Genesys project                     
                    'use client';

export interface IFase {
// 202501251
    id: number;
 
	justica: number,
	area: number,
	descricao: string,
}

export interface IFaseFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IFaseIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}